+++ source/package/adaptation_pathways/graph/node/action_begin.py
@@ -3,6 +3,8 @@ from .node import Node
 
 
 class ActionBegin(Node):
+    _action: Action
+
     def __init__(self, action: Action) -> None:
         super().__init__(f"[{action.name}")
         self._action = action